backend/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.2
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.118.3
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.0.0
pydantic_core==2.41.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
email-validator==2.1.0
# Resume Analyzer dependencies
google-generativeai==0.8.3
pypdf==5.1.0
pdfminer.six==20240706
tenacity==9.0.0

